fix: Set QQ Group Code to type Long

+++ src/main/kotlin/tech/ixor/routes/controller/qqbot/QQBotRequests.kt
@@ -2,6 +2,6 @@ package tech.ixor.routes.controller.qqbot
 
 data class QQBotSendMessageRequest(
     val host: String, val port: Int, val authKey: String,
-    val groupCode: Int, val source: String,
+    val groupCode: Long, val source: String,
     val sender: String, val message: String
 )
